+++ CHANGELOG.md
@@ -85,6 +85,7 @@ Changelog
  * renamed `ManagerPlugin` to `PuliPlugin`
  * changed `PuliPlugin::activate()` to receive the `Puli` instance
  * renamed `Package::getLoadError()` to `getLoadErrors()`
+ * added `Config::getBaseConfig()`
  
 
 * 1.0.0-beta (2015-01-12)